+++ options.go
@@ -1,6 +1,6 @@
 package smt
 
-// Option is a function that configures SMT.
+// Option is a function that configures SparseMerkleTree.
 type Option func(*BaseSMT)
 
 func SetPathHasher(ph PathHasher) Option {
